challenge-2/consumer/requirements.txt
kafka-python
oauth2client
google-cloud-bigquery